Added DES project
Created a DES implementation which is used to explore the avalanche effect in data encryption.  Does this by analyzing the resulting bit strings of 1-bit permutations of different encipherments at each stage of enchiperment.  Results are averaged and outputted for both Key and Plaintext permutations.  In addition, there is a comparison with edited round functions from DES.